Header: DTO PK 자료형 수정 및 PostRepository 생성 Body: POST 관련 DTO PK Long 자료형으로 수정, PostRepository 인터페이스 생성 Footer:

+++ src/main/java/com/ssafy/eoot/post/dto/PostImages.java
@@ -13,7 +13,7 @@ import java.sql.Timestamp;
 public class PostImages {
     // TODO : 이미지 타입 필드 회의 완료 후 추가
     @Id
-    private int imageId;
+    private Long imageId;
 
     @NotBlank
     private String originName;
